+++ src/main/kotlin/lotto/utils/RateCalculator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.utils
 
 import java.math.BigDecimal
 import java.math.RoundingMode
